Remove unnecessary custom jsmapstr type. It can be used as a regular JS map since its only purpose is to be read.
