+++ x-pack/filebeat/input/awss3/script_session.go
@@ -56,7 +56,7 @@ func newSession(p *goja.Program, conf scriptConfig, test bool) (*session, error)
 	)
 
 	// Register constructors for 'new XXXX' to enable creating them from the JS code.
-	s.vm.Set("MapStr", newJSMapStrConstructor(s))
+	// s.vm.Set("MapStr", newJSMapStrConstructor(s))
 	s.vm.Set("S3EventV2", newJSS3EventV2Constructor(s))
 
 	if _, err := s.vm.RunProgram(p); err != nil {
@@ -159,12 +159,12 @@ func (s *session) runParseFunc(m common.MapStr) (out []s3EventV2, err error) {
 		defer t.Stop()
 	}
 
-	jsm, err := newJSMapStr(s, m)
-	if err != nil {
-		return nil, fmt.Errorf("failed initializing JS object from map: %w", err)
-	}
+	// jsm, err := newJSMapStr(s, m)
+	// if err != nil {
+	// 	return nil, fmt.Errorf("failed initializing JS object from map: %w", err)
+	// }
 
-	v, err := s.parseFunc(goja.Undefined(), jsm.obj)
+	v, err := s.parseFunc(goja.Undefined(), s.vm.ToValue(m))
 	if err != nil {
 		return nil, fmt.Errorf("failed in parse function: %w", err)
 	}
